Revert "Reader iterator unification (#12387)"
This reverts commit ed0816f73bac5477da1cd5b92ff3d27fa6b8ad1d.
