+++ ydb/core/tx/columnshard/engines/reader/abstract/read_metadata.h
@@ -30,7 +30,7 @@ public:
 };
 
 // Holds all metadata that is needed to perform read/scan
-class TReadMetadataBase {
+struct TReadMetadataBase {
 public:
     enum class ESorting {
         NONE = 0 /* "not_sorted" */,
@@ -149,8 +149,8 @@ public:
 
     ui64 Limit = 0;
 
-    virtual TString DebugString() const {
-        return TStringBuilder() << " predicate{" << (PKRangesFilter ? PKRangesFilter->DebugString() : "no_initialized") << "}"
+    virtual void Dump(IOutputStream& out) const {
+        out << " predicate{" << (PKRangesFilter ? PKRangesFilter->DebugString() : "no_initialized") << "}"
             << " " << Sorting << " sorted";
     }
 
@@ -174,6 +174,12 @@ public:
 
     virtual std::unique_ptr<TScanIteratorBase> StartScan(const std::shared_ptr<TReadContext>& readContext) const = 0;
     virtual std::vector<TNameTypeInfo> GetKeyYqlSchema() const = 0;
+
+    // TODO:  can this only be done for base class?
+    friend IOutputStream& operator<<(IOutputStream& out, const TReadMetadataBase& meta) {
+        meta.Dump(out);
+        return out;
+    }
 
     const TProgramContainer& GetProgram() const {
         return Program;
