+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/ya.make
@@ -3,8 +3,6 @@ LIBRARY()
 SRCS(
     fetched_data.cpp
     columns_set.cpp
-    iterator.cpp
-    context.cpp
 )
 
 PEERDIR(
