+++ ydb/core/tx/columnshard/engines/reader/common_reader/ya.make
@@ -5,7 +5,6 @@ SRCS(
 
 PEERDIR(
     ydb/core/tx/columnshard/engines/reader/common_reader/iterator
-    ydb/core/tx/columnshard/engines/reader/common_reader/constructor
 )
 
 END()
