+++ ydb/core/tx/columnshard/engines/reader/simple_reader/constructor/ya.make
@@ -8,7 +8,6 @@ SRCS(
 
 PEERDIR(
     ydb/core/tx/columnshard/engines/reader/abstract
-    ydb/core/tx/columnshard/engines/reader/common_reader/constructor
     ydb/core/kqp/compute_actor
 )
