+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/iterator.h
@@ -1,15 +1,103 @@
 #pragma once
-#include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/iterator.h>
+#include <ydb/core/tx/columnshard/counters/scan.h>
+#include <ydb/core/tx/columnshard/engines/reader/abstract/abstract.h>
+#include <ydb/core/tx/columnshard/engines/reader/abstract/read_context.h>
 
 namespace NKikimr::NOlap::NReader::NSimple {
 
-class TColumnShardScanIterator: public NCommon::TColumnShardScanIterator {
-private:
-    using TBase = NCommon::TColumnShardScanIterator;
-    virtual void FillReadyResults() override;
+struct TReadMetadata;
 
+class TReadyResults {
+private:
+    const NColumnShard::TConcreteScanCounters Counters;
+    std::deque<std::shared_ptr<TPartialReadResult>> Data;
+    i64 RecordsCount = 0;
 public:
-    using TBase::TBase;
+    TString DebugString() const {
+        TStringBuilder sb;
+        sb
+            << "count:" << Data.size() << ";"
+            << "records_count:" << RecordsCount << ";"
+            ;
+        if (Data.size()) {
+            sb << "schema=" << Data.front()->GetResultBatch().schema()->ToString() << ";";
+        }
+        return sb;
+    }
+    TReadyResults(const NColumnShard::TConcreteScanCounters& counters)
+        : Counters(counters)
+    {
+
+    }
+    const std::shared_ptr<TPartialReadResult>& emplace_back(std::shared_ptr<TPartialReadResult>&& v) {
+        AFL_VERIFY(!!v);
+        RecordsCount += v->GetResultBatch().num_rows();
+        Data.emplace_back(std::move(v));
+        return Data.back();
+    }
+    std::shared_ptr<TPartialReadResult> pop_front() {
+        if (Data.empty()) {
+            return {};
+        }
+        auto result = std::move(Data.front());
+        AFL_VERIFY(RecordsCount >= result->GetResultBatch().num_rows());
+        RecordsCount -= result->GetResultBatch().num_rows();
+        Data.pop_front();
+        return result;
+    }
+    bool empty() const {
+        return Data.empty();
+    }
+    size_t size() const {
+        return Data.size();
+    }
 };
 
-}   // namespace NKikimr::NOlap::NReader::NSimple
+class TColumnShardScanIterator: public TScanIteratorBase {
+private:
+    std::shared_ptr<TReadContext> Context;
+    std::shared_ptr<const TReadMetadata> ReadMetadata;
+    TReadyResults ReadyResults;
+    std::shared_ptr<IDataReader> IndexedData;
+    ui64 ItemsRead = 0;
+    const i64 MaxRowsInBatch = 5000;
+    virtual void DoOnSentDataFromInterval(const ui32 intervalIdx) const override;
+
+public:
+    TColumnShardScanIterator(const std::shared_ptr<TReadContext>& context, const std::shared_ptr<const TReadMetadata>& readMetadata);
+    ~TColumnShardScanIterator();
+
+    virtual TConclusionStatus Start() override {
+        AFL_VERIFY(IndexedData);
+        return IndexedData->Start();
+    }
+
+    virtual std::optional<ui32> GetAvailableResultsCount() const override {
+        return ReadyResults.size();
+    }
+
+    virtual const TReadStats& GetStats() const override;
+
+    virtual TString DebugString(const bool verbose) const override {
+        return TStringBuilder()
+            << "ready_results:(" << ReadyResults.DebugString() << ");"
+            << "indexed_data:(" << IndexedData->DebugString(verbose) << ")"
+            ;
+    }
+
+    virtual void Apply(const std::shared_ptr<IApplyAction>& task) override;
+
+    bool Finished() const  override {
+        return IndexedData->IsFinished() && ReadyResults.empty();
+    }
+
+    virtual TConclusion<std::shared_ptr<TPartialReadResult>> GetBatch() override;
+    virtual void PrepareResults() override;
+
+    virtual TConclusion<bool> ReadNextInterval() override;
+
+private:
+    void FillReadyResults();
+};
+
+}
