+++ ydb/core/tx/columnshard/engines/reader/transaction/tx_scan.cpp
@@ -139,7 +139,7 @@ void TTxScan::Complete(const TActorContext& ctx) {
 
     TStringBuilder detailedInfo;
     if (IS_LOG_PRIORITY_ENABLED(NActors::NLog::PRI_TRACE, NKikimrServices::TX_COLUMNSHARD)) {
-        detailedInfo << " read metadata: (" << readMetadataRange->DebugString() << ")"
+        detailedInfo << " read metadata: (" << *readMetadataRange << ")"
                      << " req: " << request;
     }
